tools/solrIndexer/requirements.txt
pdftotext==2.1.4
python-dateutil==2.8.1
requests==2.23.0
tqdm==4.45.0